+++ functions/src/index.ts
@@ -5,7 +5,7 @@ import tagitemsinlist from './Scrapers/tagItemsInList';
 import updatecreatedandeditedatmillis from './updateCreatedAndEditedAtMillis';
 import uploadToSubsplash from './uploadToSubsplash';
 import setUserRole from './setUserRole';
-import addintrooutro from './addIntroOutro';
+import addintrooutrotaskhandler from './addintrooutrotaskhandler';
 import deleteFromSubsplash from './deleteFromSubsplash';
 import uploadimage from './handleImageUpload';
 import saveimage from './saveImage';
@@ -38,7 +38,7 @@ exports.editSubsplashSermon = editSubsplashSermon;
 exports.editSoundCloudSermon = editSoundCloudSermon;
 exports.deleteFromSubsplash = deleteFromSubsplash;
 exports.setUserRole = setUserRole;
-exports.addintrooutro = addintrooutro;
+exports.addintrooutrotaskhandler = addintrooutrotaskhandler;
 exports.uploadimage = uploadimage;
 exports.getimage = getimage;
 exports.listusers = listusers;
